relocate make_available_xml_midi_positions
