+++ data_class.py
@@ -236,6 +236,8 @@ class PieceData:
     def _align_perform_with_score(self, perform):
         perform.match_between_xml_perf = matching.match_score_pair2perform(self.score_pairs, perform.midi_notes, perform.corresp)
         perform.pairs = matching.make_xml_midi_pair(self.xml_notes, perform.midi_notes, perform.match_between_xml_perf)
+        perform.pairs, perform.valid_position_pair = make_available_xml_midi_positions(pairs)
+
         print('Performance path is ', perform.midi_path)
         perform._count_matched_notes()
