+++ xml_midi_matching.py
@@ -1,4 +1,5 @@
 import csv
+import copy
 
 
 def match_xml_to_midi(xml_notes, midi_notes):
@@ -139,3 +140,163 @@ def find_by_attr(alist, value1, value2):
         if abs(obj.start - value1) < 0.02 and obj.pitch == value2:
             return i
     return []
+
+
+def make_available_xml_midi_positions(pairs):
+    # global NUM_EXCLUDED_NOTES
+    class PositionPair:
+        def __init__(self, xml_pos, time, pitch, index, divisions):
+            self.xml_position = xml_pos
+            self.time_position = time
+            self.pitch = pitch
+            self.index = index
+            self.divisions = divisions
+            self.is_arpeggiate = False
+
+    available_pairs = []
+    num_pairs = len(pairs)
+    for i in range(num_pairs):
+        pair = pairs[i]
+        if not pair == []:
+            xml_note = pair['xml']
+            midi_note = pair['midi']
+            xml_pos = xml_note.note_duration.xml_position
+            time = midi_note.start
+            divisions = xml_note.state_fixed.divisions
+            if not xml_note.note_duration.is_grace_note:
+                pos_pair = PositionPair(xml_pos, time, xml_note.pitch[1], i, divisions)
+                if xml_note.note_notations.is_arpeggiate:
+                    pos_pair.is_arpeggiate = True
+                available_pairs.append(pos_pair)
+
+    # available_pairs = save_lowest_note_on_same_position(available_pairs)
+    available_pairs, mismatched_indexes = make_average_onset_cleaned_pair(available_pairs)
+    print('Number of mismatched notes: ', len(mismatched_indexes))
+    # NUM_EXCLUDED_NOTES += len(mismatched_indexes)
+    for index in mismatched_indexes:
+        pairs[index] = []
+
+    return pairs, available_pairs
+
+
+def make_average_onset_cleaned_pair(position_pairs, maximum_qpm=600):
+    length = len(position_pairs)
+    previous_position = -float("Inf")
+    previous_time = -float("Inf")
+    previous_index = 0
+    # position_pairs.sort(key=lambda x: (x.xml_position, x.pitch))
+    cleaned_list = list()
+    notes_in_chord = list()
+    mismatched_indexes = list()
+    for i in range(length):
+        pos_pair = position_pairs[i]
+        current_position = pos_pair.xml_position
+        current_time = pos_pair.time_position
+        if current_position > previous_position >= 0:
+            minimum_time_interval = (current_position - previous_position) / pos_pair.divisions / maximum_qpm * 60 + 0.01
+        else:
+            minimum_time_interval = 0
+        if current_position > previous_position and current_time > previous_time + minimum_time_interval:
+            if len(notes_in_chord) > 0:
+                average_pos_pair = copy.copy(notes_in_chord[0])
+                notes_in_chord_cleaned, average_pos_pair.time_position = get_average_onset_time(notes_in_chord)
+                if len(cleaned_list) == 0 or average_pos_pair.time_position > cleaned_list[-1].time_position + (
+                        (average_pos_pair.xml_position - cleaned_list[-1].xml_position) /
+                        average_pos_pair.divisions / maximum_qpm * 60 + 0.01):
+                    cleaned_list.append(average_pos_pair)
+                    for note in notes_in_chord:
+                        if note not in notes_in_chord_cleaned:
+                            # print('the note is far from other notes in the chord')
+                            mismatched_indexes.append(note.index)
+                else:
+                    # print('the onset is too close to the previous onset', average_pos_pair.xml_position, cleaned_list[-1].xml_position, average_pos_pair.time_position, cleaned_list[-1].time_position)
+                    for note in notes_in_chord:
+                        mismatched_indexes.append(note.index)
+            notes_in_chord = list()
+            notes_in_chord.append(pos_pair)
+            previous_position = current_position
+            previous_time = current_time
+            previous_index = i
+        elif current_position == previous_position:
+            notes_in_chord.append(pos_pair)
+        else:
+            # print('the note is too close to the previous note', current_position - previous_position, current_time - previous_time)
+            # print(previous_position, current_position, previous_time, current_time)
+            mismatched_indexes.append(position_pairs[previous_index].index)
+            mismatched_indexes.append(pos_pair.index)
+
+    return cleaned_list, mismatched_indexes
+
+
+
+def make_available_note_feature_list(notes, features, predicted):
+    class PosTempoPair:
+        def __init__(self, xml_pos, pitch, qpm, index, divisions, time_pos):
+            self.xml_position = xml_pos
+            self.qpm = qpm
+            self.index = index
+            self.divisions = divisions
+            self.pitch = pitch
+            self.time_position = time_pos
+            self.is_arpeggiate = False
+
+    if not predicted:
+        available_notes = []
+        num_features = len(features)
+        for i in range(num_features):
+            feature = features[i]
+            if not feature.qpm == None:
+                xml_note = notes[i]
+                xml_pos = xml_note.note_duration.xml_position
+                time_pos = feature.midi_start
+                divisions = xml_note.state_fixed.divisions
+                qpm = feature.qpm
+                pos_pair = PosTempoPair(xml_pos, xml_note.pitch[1], qpm, i, divisions, time_pos)
+                if xml_note.note_notations.is_arpeggiate:
+                    pos_pair.is_arpeggiate = True
+                available_notes.append(pos_pair)
+
+    else:
+        available_notes = []
+        num_features = len(features)
+        for i in range(num_features):
+            feature = features[i]
+            xml_note = notes[i]
+            xml_pos = xml_note.note_duration.xml_position
+            time_pos = xml_note.note_duration.time_position
+            divisions = xml_note.state_fixed.divisions
+            qpm = feature.qpm
+            pos_pair = PosTempoPair(xml_pos, xml_note.pitch[1], qpm, i, divisions, time_pos)
+            available_notes.append(pos_pair)
+
+    # minimum_time_interval = 0.05
+    # available_notes = save_lowest_note_on_same_position(available_notes, minimum_time_interval)
+    available_notes, _ = make_average_onset_cleaned_pair(available_notes)
+    return available_notes
+
+
+def get_average_onset_time(notes_in_chord_saved, threshold=0.2):
+    # notes_in_chord: list of PosTempoPair, len > 0
+    notes_in_chord = copy.copy(notes_in_chord_saved)
+    average_onset_time = 0
+    for pos_pair in notes_in_chord:
+        average_onset_time += pos_pair.time_position
+        if pos_pair.is_arpeggiate:
+            threshold = 1
+    average_onset_time /= len(notes_in_chord)
+
+    # check whether there is mis-matched note
+    deviations = list()
+    for pos_pair in notes_in_chord:
+        dev = abs(pos_pair.time_position - average_onset_time)
+        deviations.append(dev)
+    if max(deviations) > threshold:
+        # print(deviations)
+        if len(notes_in_chord) == 2:
+            del notes_in_chord[0:2]
+        else:
+            index = deviations.index(max(deviations))
+            del notes_in_chord[index]
+            notes_in_chord, average_onset_time = get_average_onset_time(notes_in_chord, threshold)
+
+    return notes_in_chord, average_onset_time
